+++ changelog.md
@@ -3,7 +3,9 @@
 * Form implement its own submit method
 * Fields, simple object to build a "form" template (can't be simpler)
 * new widgets !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
+* bye bye : InputText,TextArea,Checkbox,Slider,RadioButtons,SelectButtons,Select
+* welcome : Input,Range,Checkbox,Radio,SelectButtons,TabsHeader,Select,Textarea
+* theses new ones are better, in all way ...
 
 ### 0.1.2  2022/03/24